Time: 0 ms (100.00%), Space: 43.9 MB (48.18%) - LeetHub

+++ 1232-check-if-it-is-a-straight-line/1232-check-if-it-is-a-straight-line.java
@@ -2,7 +2,7 @@ class Solution {
     public boolean checkStraightLine(int[][] coordinates) {
         if(coordinates.length == 2) return true;
         for(int i = 1; i < coordinates.length - 1; i++){
-            if(((double)coordinates[i][1] - (double)coordinates[i-1][1]) * ((double)coordinates[i + 1][0] - (double)coordinates[i][0]) !=  ((double)coordinates[i + 1][1] - (double)coordinates[i][1]) * ((double)coordinates[i][0] - (double)coordinates[i-1][0])) return false;
+            if((coordinates[i][1] - coordinates[i-1][1]) * (coordinates[i + 1][0] - coordinates[i][0]) !=  (coordinates[i + 1][1] - coordinates[i][1]) * (coordinates[i][0] - coordinates[i-1][0])) return false;
         }
         return true;
     }
